+++ src/effects/ClickRemoval.h
@@ -61,7 +61,7 @@ private:
    bool ProcessOne(int count, WaveTrack * track,
                    sampleCount start, sampleCount len);
 
-   bool RemoveClicks(int len, float *buffer);
+   bool RemoveClicks(size_t len, float *buffer);
 
    void OnWidthText(wxCommandEvent & evt);
    void OnThreshText(wxCommandEvent & evt);
